Rename folder common-components to components
